+++ packages/expo-camera/src/CameraModule/CameraModule.ts
@@ -2,30 +2,20 @@
 import * as React from 'react';
 
 import { CameraPictureOptions } from '../Camera.types';
-import { CameraType, CapturedPicture, CaptureOptions, ImageType } from './CameraModule.types';
+import {
+  CameraType,
+  CapturedPicture,
+  WebCameraSettings,
+  CaptureOptions,
+  ImageType,
+} from './CameraModule.types';
 import * as Utils from './CameraUtils';
-import * as CapabilityUtils from './CapabilityUtils';
 import { FacingModeToCameraType } from './constants';
 
 export { ImageType, CameraType, CaptureOptions };
 
 type OnCameraReadyListener = () => void;
 type OnMountErrorListener = (event: { nativeEvent: Error }) => void;
-
-export type WebCameraSettings = Partial<{
-  autoFocus: string;
-  flashMode: string;
-  whiteBalance: string;
-  exposureCompensation: number;
-  colorTemperature: number;
-  iso: number;
-  brightness: number;
-  contrast: number;
-  saturation: number;
-  sharpness: number;
-  focusDistance: number;
-  zoom: number;
-}>;
 
 const VALID_SETTINGS_KEYS = [
   'autoFocus',
@@ -74,7 +64,7 @@ export function useCameraStream(
   }: { onCameraReady?: OnCameraReadyListener; onMountError?: OnMountErrorListener }
 ): {
   type: CameraType | null;
-  resume: () => Promise<MediaStream | null>;
+  resume: () => Promise<void>;
   stop: () => void;
   capture: (config: CameraPictureOptions) => CapturedPicture;
 } {
@@ -90,7 +80,7 @@ export function useCameraStream(
   const [realType, setRealType] = React.useState<CameraType | null>(null);
 
   useLoadedVideo(video, () => {
-    syncTrackCapabilities(type, stream.current, capabilities.current);
+    Utils.syncTrackCapabilities(type, stream.current, capabilities.current);
   });
 
   React.useEffect(() => {
@@ -121,7 +111,7 @@ export function useCameraStream(
 
     const nextWebCameraSettings = { ...capabilities.current, ...changes };
     if (hasChanges) {
-      syncTrackCapabilities(type, stream.current, changes);
+      Utils.syncTrackCapabilities(type, stream.current, changes);
     }
 
     capabilities.current = nextWebCameraSettings;
@@ -156,12 +146,12 @@ export function useCameraStream(
     if (!video.current) {
       return;
     }
-    setVideoSource(video.current, stream.current);
+    Utils.setVideoSource(video.current, stream.current);
   }, [stream.current, video.current]);
 
   const stopAsync = (): void => {
     console.log('stop', stream.current);
-    stopMediaStream(stream.current);
+    Utils.stopMediaStream(stream.current);
     stream.current = null;
   };
 
@@ -180,7 +170,7 @@ export function useCameraStream(
       onMountError?.({ nativeEvent: error });
       return;
     }
-    if (compareStreams(streamDevice, stream.current)) {
+    if (Utils.compareStreams(streamDevice, stream.current)) {
       // Do nothing if the streams are the same.
       // This happens when the device only supports one camera (i.e. desktop) and the mode was toggled between front/back while already active.
       // Without this check there is a screen flash while the video switches.
@@ -198,203 +188,7 @@ export function useCameraStream(
     resume: resumePreview,
     stop: stopAsync,
     capture(config: CameraPictureOptions): CapturedPicture {
-      return capture(video.current!, streamSettings.current!, config);
+      return Utils.capture(video.current!, streamSettings.current!, config);
     },
   };
-}
-
-function compareStreams(a: MediaStream | null, b: MediaStream | null): boolean {
-  if (!a || !b) return false;
-  const settingsA = a.getTracks()[0].getSettings();
-  const settingsB = b.getTracks()[0].getSettings();
-  return settingsA.deviceId === settingsB.deviceId;
-}
-
-function capture(
-  video: HTMLVideoElement,
-  settings: MediaTrackSettings,
-  config: CameraPictureOptions
-): CapturedPicture {
-  const base64 = Utils.captureImage(video, config);
-
-  const capturedPicture: CapturedPicture = {
-    uri: base64,
-    base64,
-    width: 0,
-    height: 0,
-  };
-
-  if (settings) {
-    const { width = 0, height = 0 } = settings;
-    capturedPicture.width = width;
-    capturedPicture.height = height;
-    // TODO: Bacon: verify/enforce exif shape.
-    capturedPicture.exif = settings;
-  }
-
-  if (config.onPictureSaved) {
-    config.onPictureSaved({ nativeEvent: { data: capturedPicture, id: config.id } });
-  }
-  return capturedPicture;
-}
-
-async function syncTrackCapabilities(
-  cameraType: CameraType,
-  stream: MediaStream | null,
-  settings: WebCameraSettings = {}
-): Promise<void> {
-  if (stream?.getVideoTracks) {
-    await Promise.all(
-      stream.getVideoTracks().map(track => onCapabilitiesReady(cameraType, track, settings))
-    );
-  }
-}
-
-// https://developer.mozilla.org/en-US/docs/Web/API/MediaTrackConstraints
-async function onCapabilitiesReady(
-  cameraType: CameraType,
-  track: MediaStreamTrack,
-  settings: WebCameraSettings = {}
-): Promise<void> {
-  console.log('sync: ', cameraType, settings);
-  const capabilities = track.getCapabilities();
-
-  // Create an empty object because if you set a constraint that isn't available an error will be thrown.
-  const constraints: MediaTrackConstraintSet = {};
-
-  // TODO: Bacon: Add `pointsOfInterest` support
-  const clampedValues = [
-    'exposureCompensation',
-    'colorTemperature',
-    'iso',
-    'brightness',
-    'contrast',
-    'saturation',
-    'sharpness',
-    'focusDistance',
-    'zoom',
-  ];
-
-  for (const property of clampedValues) {
-    if (capabilities[property]) {
-      constraints[property] = convertNormalizedSetting(capabilities[property], settings[property]);
-    }
-  }
-
-  const _validatedConstrainedValue = (key, propName, converter) =>
-    validatedConstrainedValue(
-      key,
-      propName,
-      converter(settings[propName]),
-      capabilities,
-      settings,
-      cameraType
-    );
-
-  if (capabilities.focusMode && settings.autoFocus !== undefined) {
-    constraints.focusMode = _validatedConstrainedValue(
-      'focusMode',
-      'autoFocus',
-      CapabilityUtils.convertAutoFocusJSONToNative
-    );
-  }
-
-  console.log('can use torch: ', capabilities.torch);
-
-  if (capabilities.torch && settings.flashMode !== undefined) {
-    constraints.torch = _validatedConstrainedValue(
-      'torch',
-      'flashMode',
-      CapabilityUtils.convertFlashModeJSONToNative
-    );
-  }
-
-  if (capabilities.whiteBalanceMode && settings.whiteBalance !== undefined) {
-    constraints.whiteBalanceMode = _validatedConstrainedValue(
-      'whiteBalanceMode',
-      'whiteBalance',
-      CapabilityUtils.convertWhiteBalanceJSONToNative
-    );
-  }
-
-  console.log('apply constraints: ', constraints);
-  await track.applyConstraints({ advanced: [constraints] as any });
-}
-
-function stopMediaStream(stream: MediaStream | null) {
-  if (!stream) return;
-  if (stream.getAudioTracks) stream.getAudioTracks().forEach(track => track.stop());
-  if (stream.getVideoTracks) stream.getVideoTracks().forEach(track => track.stop());
-  if (isMediaStreamTrack(stream)) stream.stop();
-}
-
-function setVideoSource(
-  video: HTMLVideoElement,
-  stream: MediaStream | MediaSource | Blob | null
-): void {
-  try {
-    video.srcObject = stream;
-  } catch {
-    if (stream) {
-      video.src = window.URL.createObjectURL(stream);
-    } else if (typeof video.src === 'string') {
-      window.URL.revokeObjectURL(video.src);
-    }
-  }
-}
-
-function isCapabilityAvailable(video: HTMLVideoElement, keyName: string): boolean {
-  const stream = video.srcObject;
-
-  if (stream instanceof MediaStream) {
-    const videoTrack = stream.getVideoTracks()[0];
-
-    if (typeof videoTrack.getCapabilities === 'undefined') {
-      return false;
-    }
-
-    const capabilities: any = videoTrack.getCapabilities();
-
-    return !!capabilities[keyName];
-  }
-
-  return false;
-}
-
-function isMediaStreamTrack(input: any): input is MediaStreamTrack {
-  return typeof input.stop === 'function';
-}
-
-function convertNormalizedSetting(range: MediaSettingsRange, value?: number): number | undefined {
-  if (!value) return;
-  // convert the normalized incoming setting to the native camera zoom range
-  const converted = convertRange(value, [range.min, range.max]);
-  // clamp value so we don't get an error
-  return Math.min(range.max, Math.max(range.min, converted));
-}
-
-function convertRange(value: number, r2: number[], r1: number[] = [0, 1]): number {
-  return ((value - r1[0]) * (r2[1] - r2[0])) / (r1[1] - r1[0]) + r2[0];
-}
-
-function validatedConstrainedValue(
-  constraintKey: string,
-  settingsKey: string,
-  convertedSetting: any,
-  capabilities: MediaTrackCapabilities,
-  settings: any,
-  cameraType: string
-): any {
-  const setting = settings[settingsKey];
-  if (
-    Array.isArray(capabilities[constraintKey]) &&
-    convertedSetting &&
-    !capabilities[constraintKey].includes(convertedSetting)
-  ) {
-    console.warn(
-      ` { ${settingsKey}: "${setting}" } (converted to "${convertedSetting}" in the browser) is not supported for camera type "${cameraType}" in your browser. Using the default value instead.`
-    );
-    return undefined;
-  }
-  return convertedSetting;
 }
